refactor: move use-character-details to client
