+++ src/app/(client)/_hooks/use-character-details.ts
